Overhaul Web API surface
Rather than returning multiple content types and redirecting to the full blob for plain-text, simply return JSON with a preview of the expansion and set the Content-Location header to the full blob. This also fixes a CORS issue because a automatic redirect in a CORS context does not include the Origin header.
In reading the blob to generate the preview, it fetches only a kilobyte.