Update state domain type references only within state functions
Due to js reference equality, updating a ref of one of the domain types in state cause a reactive change, even if the two objects have the same logical/value equality. This means needing to have numerous spaghetti-like checks for value equality of the domain types within the components.

This update centralizes all mutation into functions provided by the state object, and these only replace the domain type references if they are not logically equivalent.