Merge the API and website repos
Previously the C#-based Web API was in a separate repo from the front-end. I first used get-filter-repo on each to move the projects into distinct directories. I then rebased the front-end on top of the API, so it will appear in history as if the two parts were always part of the same repo. This commit handles merging the parts that ought not be separate.