Warm-up decimal expansions API upon first visting the melodies page
I currently have the API portion running in a container app that may scale to zero, and the initial query on the OEIS side is slow. This can create incedible cold-start delays. As a work-around, the melodies page will call the get random expansion endpoint to ensure that the instance is running and that the OEIS system has processed the query (based on my assumption that they somehow cache query restults/plans for a time).