[API] Pivot to consumption of OEIS search result text format
The JSON format suddenly changed without warning. It no longer returns an object containing the index and total count with an array of the search results, instead returning the array directly. This makes it impossible to page the results are seek a specific result because I don’t know the total count. The text format still includes a header with the ordinal and total count.